Clean Code, String To constant , Image loader to Extension Fun
