+++ app/src/main/java/com/ameri/newsapi/ui/home/SportSection.kt
@@ -1,16 +1,13 @@
 package com.ameri.newsapi.ui.home
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -24,6 +21,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.ameri.newsapi.data.models.ResponseData
+import com.ameri.newsapi.ui.component.LoadingView
 import com.ameri.newsapi.util.constant.SPORTS
 import com.ameri.newsapi.util.network.NetworkRequest
 import com.ameri.newsapi.viewmodel.HomeViewModel
@@ -61,20 +59,7 @@ fun SportSection(viewModel: HomeViewModel? = hiltViewModel()) {
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         if (loading) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(20.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center
-            ) {
-                CircularProgressIndicator(
-                    modifier = Modifier
-                        .height(34.dp)
-                        .width(34.dp),
-                    color = MaterialTheme.colorScheme.tertiary
-                )
-            }
+            LoadingView()
         } else {
             Column(
                 modifier = Modifier
